### 2025-05-05 #283 @editorialbot
[PRE REVIEW]: nemogb: An R Package for Calculating Course Grades
**Submitting author:** @nikita-jaya (<a href="http://orcid.org/0009-0001-2708-0210">Nikita Jayaprakash</a>)
**Repository:** https://github.com/nemogb-dev/nemogb-r/
**Branch with paper.md** (empty if default branch): jose-paper
**Version:** v1.1.0
