### 2025-07-06 #283 @labarba
hi 👋  – thank you for your patience ... we're facing a backlog of more than two dozen submissions awaiting to be assigned an editor, and an all-volunteer board that does the best they can...
