### 2025-07-30 #283 @bleds22e
@lwjohnst86 Would you be interested in serving as a reviewer for this JOSE paper and R package?
