### 2025-07-31 #283 @labarba
Way to go, everybody! Please also welcome @bleds22e (Ellen), our new member of the editorial board. This will be her first editorial task, and I am here to help in any way necessary!
