### 2025-08-28 #303 @editorialbot
[REVIEW]: nemogb: An R Package for Calculating Course Grades
**Submitting author:** @nikita-jaya (<a href="http://orcid.org/0009-0001-2708-0210">Nikita Jayaprakash</a>)
**Repository:** https://github.com/nemogb-dev/nemogb-r/
**Branch with paper.md** (empty if default branch): main
**Version:** v1.2.2
**Paper kind:** software
