### 2025-08-28 #303 @lwjohnst86
- [x] **Version:** Does the release version given match the GitHub release?
