### 2025-09-10 #303 @lwjohnst86
@nikita-jaya the version listed in this review is older than the current version (1.1.0 vs 1.2.0). I think JOSE wants them to be the same (see my checklist).
