### 2025-09-10 nemogb-dev/nemogb-r#110 @lwjohnst86
Include an example Gradescope data and policy file to use in examples
Related to review: https://github.com/openjournals/jose-reviews/issues/303

The example usage doc in the README implies I have a `gs_demo.csv` file on my computer, but I don't. It would be very helpful to actually include that in this package, as an exposed function that links to an demo file that comes with the package (see using `fs::path_package()` to provide link to a package's bundled files). This also applies to the policy file.

I see there are some of these objects bundled in the package, but they aren't described in the README.
