### 2025-09-10 nemogb-dev/nemogb-r#111 @lwjohnst86
Image in `working-with-canvas-data` vignette doesn't render on the website
Related to review: https://github.com/openjournals/jose-reviews/issues/303

In the section "getting the data". It seems to be pointing to a temp location when inspecting the HTML.
