### 2025-09-10 nemogb-dev/nemogb-r#112 @lwjohnst86
Provide `paper.md` somewhere that is accessible or obvious
Related to review: https://github.com/openjournals/jose-reviews/issues/303

Trying to review the paper, but of course, it isn't in the main branch. Please put it somewhere that is easily accessible and review-able (not a branch). I believe JOSE has some guidelines on this.
