### 2025-09-17 #303 @nikita-jaya
> [@nikita-jaya](https://github.com/nikita-jaya) the version listed in this review is older than the current version (1.1.0 vs 1.2.0). I think JOSE wants them to be the same (see my checklist).

@lwjohnst86 thank you for the comments so far. I'll be addressing these soon.
