### 2025-09-17 #303 @nikita-jaya
@bleds22e we had made some small developments shortly after submitting to JOSE and released a new Github version. Would you be able to update our review version to v1.2.0?
