### 2025-09-17 a2f4806
Removing image
Pretty sure we wanted to remove all images, we must have missed this one
