+++ pkgdown.yml
@@ -1,5 +1,5 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   complex-syllabus: complex-syllabus.html
@@ -10,7 +10,7 @@ articles:
   simple-syllabus: simple-syllabus.html
   slip-days: slip-days.html
   working-with-canvas-data: working-with-canvas-data.html
-last_built: 2025-03-19T16:54Z
+last_built: 2025-10-01T15:58Z
 urls:
   reference: https://github.com/nemogb-dev/nemogb-r/reference
   article: https://github.com/nemogb-dev/nemogb-r/articles
