### 2025-10-01 c411df9
add paper into main-temp
