+++ paper/paper.md
@@ -1,0 +1,99 @@
+---
+title: 'NemoGB: An R Package for Calculating Course Grades'
+tags:
+  - R
+  - education
+  - data science
+authors:
+- name: Nikita Jayaprakash
+  orcid: "0009-0001-2708-0210"
+  equal-contrib: true
+  affiliation: 1
+- name: Monika Voutov
+  orcid: "0009-0000-9923-0954"
+  equal-contrib: true
+  affiliation: "1, 2"
+- name: Zach Turner
+  orcid: "0009-0005-1829-2325"
+  equal-contrib: false
+  affiliation: 1
+- name: Andrew Bray
+  orcid: "0000-0002-4037-7414"
+  equal-contrib: true
+  affiliation: 1
+affiliations:
+  - index: 1
+    name: UC Berkeley, Department of Statistics
+  - index: 2
+    name: UC Berkeley, College of Engineering
+
+citation_author: Jayaprakash et al.
+date: "5 November 2024"
+year: 2024
+bibliography: paper.bib
+output: 
+  rticles::joss_article:
+    journal: JOSE
+    keep_tex: true
+journal: JOSE
+---
+
+# Summary
+
+`NemoGB` [@NemoGB-Package] allows for accurate and systematic computations of the final course letter grades. These computations require two inputs: a specifically structured YAML file representing the grading policy from the class syllabus and the assignment grades in CSV (comma-separated value) format from Gradescope [@10.1145/3051457.3051466] or other similar learning management systems, like Canvas [@canvas_gradebook_guide]. 
+The package uses these two inputs to break down any complex syllabi into methodical computations that can be documented and tested.
+
+# Statement of Need
+
+While the final grade at the end of a course is an elementary part of most college courses, the computations for these grades quickly become deceptively intricate, especially with larger STEM classes that use various complexities to accommodate a diverse student body. Even though most classes use slight variations of the same policies, many LMS cannot sustain these complex computations. In response, courses will turn to hard-coded scripts. These scripts quickly accumulate hundreds of lines of code, and there is no method to assess accuracy of the final computation. 
+
+
+
+`NemoGB` is an R package that maintains the structure and complexity of a course grade while guaranteeing accuracy through comprehensive unit-testing. The challenges of consistency and precision in grading systems are addressed by applying the practices of data analysis and the principles of software development. The rigorous unit-testing in the package minimizes computational error and reduces the manual inputs, significantly lowers the risks of typographic and logical errors in scripts. Because of this, course grades can be computed accurately and quickly: the accuracy allows course instructors to have reliable grade computations, and the speed allows them to compute grades throughout the semester in order to monitor student progress. The structure of the package -- and the open-source nature of it -- allows for courses to contribute functionality that is unique to their course. This R package also functions as the backend of the NemoGB Shiny app [@NemoGB-App], which lets the user create their grading policy file in a straightforward way. 
+
+Grading Workflow WITH NemoGB             |  Grading Workflow WITHOUT NemoGB
+:-------------------------:|:-------------------------:
+![](with_nemogb_workflow.png){height="150pt"}  |  ![](without_nemogb_workflow.png){height="150pt"}
+
+# Story of the Project
+
+As students, we understand the importance of having an informed instructor who understands how we stand in their course. This began as an independent study centered around using R Shiny in order to create a web-based grading tool that allowed for the computation of course grades for diverse syllabi. We wanted to develop a tool that allowed for flexibility in their syllabus and precision in their computations. This led to the development of this package, which operates as the backend of the NemoGB app but can also be used independently to compute course grades. 
+
+
+# Underlying Principle
+
[…]
