+++ paper/paper.bib
@@ -1,0 +1,87 @@
+@misc{Harrison_2018, 
+title={Turnitin acquires Gradescope, expands grading and Assessment Solutions for Higher Education}, 
+url={https://press.jharrisonpr.com/turnitin-acquires-gradescope-expands-grading-and-assessment-solutions-for-higher-education}, 
+journal={J Harrison Public Relations Group}, 
+author={Harrison, Jennifer}, 
+year={2018}, 
+month={Oct},
+note = {Accessed: 2024-07-19}
+} 
+
+@inproceedings{10.1145/3051457.3051466,
+author = {Singh, Arjun and Karayev, Sergey and Gutowski, Kevin and Abbeel, Pieter},
+title = {Gradescope: A Fast, Flexible, and Fair System for Scalable Assessment of Handwritten Work},
+year = {2017},
+isbn = {9781450344500},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/3051457.3051466},
+doi = {10.1145/3051457.3051466},
+abstract = {We present a system for online assessment of handwritten homework assignments and exams. First, either instructors or students scan and upload handwritten work. Instructors then grade the work and distribute the results using a web-based platform. Our system optimizes for three key dimensions: speed, consistency, and flexibility. The primary innovation enabling improvements in all three dimensions is a dynamically evolving rubric for each question on an assessment. We also describe how the system minimizes the overhead incurred in the digitization process. Our system has been in use for four years, with instructors at 200 institutions having graded over 10 million pages of student work. We present results as user-reported data and feedback regarding time saved grading, enjoyment, and student experience. Two-thirds of responders report saving 30\% or more time relative to their traditional workflow. We also find that the time spent grading an individual response to a question rapidly decays with the number of responses to that question that the grader has already graded.},
+booktitle = {Proceedings of the Fourth (2017) ACM Conference on Learning @ Scale},
+pages = {81–88},
+numpages = {8},
+keywords = {scaling large courses, rubric-based grading, learning assessment, education, computer-assisted instruction},
+location = {Cambridge, Massachusetts, USA},
+series = {L@S '17}
+}
+
+
+@misc{NemoGB-App,
+  author = {Voutov*, Monika and Jayaprakash*, Nikita and Bray, Andrew},
+  title = {NemoGB Web App},
+  year = {2024},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/nemogb-dev/nemogb-app.git}}
+}
+
+@misc{NemoGB-Package,
+  author = {Jayaprakash*, Nikita and Voutov*, Monika and Turner, Zach and Bray, Andrew},
+  title = {NemoGB Package},
+  year = {2024},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/nemogb-dev/nemogb-r.git}}
+}
+
+@misc{canvas_gradebook_guide,
+  author       = {{Canvas Community}},
+  title        = {How do I use the Gradebook?},
+  year         = {2024},
+  url          = {https://community.canvaslms.com/t5/Instructor-Guide/How-do-I-use-the-Gradebook/ta-p/701},
+  note         = {Accessed: 2024-07-22}
+}
+
+@misc{moodle_grade_calculations,
+  author       = {{Moodle Docs}},
+  title        = {Grade calculations},
+  year         = {2024},
+  url          = {https://docs.moodle.org/404/en/Grade_calculations},
+  note         = {Accessed: 2024-07-22}
+}
+
[…]
