### 2025-10-07 nemogb-dev/nemogb-r#118 @lwjohnst86
Don't use `tidytemplate` for the website theme
The tidyverse team explicitly requests *not* to use their template: https://github.com/tidyverse/tidytemplate#overview

The base theme is honestly pretty good, can you use just that?

Related to review at https://github.com/openjournals/jose-reviews/issues/303
