### 2025-10-07 nemogb-dev/nemogb-r#119 @lwjohnst86
Run styler or Air formatter to format code for easier readability
Related to review at https://github.com/openjournals/jose-reviews/issues/303

It would make it much easier to ready the code (as examples in the docs and the code in the package) if it was styled correctly. Could you use the styler or [Air](https://posit-dev.github.io/air/) formatter on the code?
