### 2025-10-07 nemogb-dev/nemogb-r#120 @lwjohnst86
Remove `.DS_Store` files from repo
Related to review at https://github.com/openjournals/jose-reviews/issues/303

These files should be ignored in Git, they shouldn't be tracked and saved. Can you remove them and update the `.gitignore` file to not track them?
