### 2025-10-07 37923b1
removing tidytemplate
