### 2025-10-07 0dc1b44
ran styler::style_pkg() on entire package
