### 2025-10-07 ad7ab66
devtools::document() updates
