+++ man/read_gs.Rd
@@ -19,7 +19,8 @@ This functions reads the Gradescope CSV file and checks for correct Gradescope
 format. Each assignment should follow the four-column format:
 \verb{Assignment Name}, \verb{Assignment Name - Max Points}, \verb{Assignment Name - Submission Time},
 \verb{Assignment Name - Lateness (H:M:S)}. All other columns are designed as ID columns.
-
+}
+\details{
 Now it is favored to use read_files.
 This function is left for backwards-compatibility.
 }
