+++ DESCRIPTION
@@ -41,7 +41,7 @@ Imports:
     stats,
     yaml
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.2
+RoxygenNote: 7.3.3
 Depends: 
     R (>= 2.10)
 LazyData: true
