+++ pkgdown.yml
@@ -10,7 +10,7 @@ articles:
   simple-syllabus: simple-syllabus.html
   slip-days: slip-days.html
   working-with-canvas-data: working-with-canvas-data.html
-last_built: 2025-10-01T23:46Z
+last_built: 2025-10-07T16:52Z
 urls:
   reference: https://github.com/nemogb-dev/nemogb-r/reference
   article: https://github.com/nemogb-dev/nemogb-r/articles
