### 2025-10-16 #303 @jbburant
- [x] **Version:** Does the release version given match the GitHub release?
