### 2025-10-16 nemogb-dev/nemogb-r#124 @jbburant
Clarify order of operations instructions
Relates to review for JOSE: https://github.com/openjournals/jose-reviews/issues/303

In the [Building a Policy File vignette](https://nemogb-dev.github.io/nemogb-r/articles/policy-files.html), you talk about the importance of order of `lateness`, `drop_n_lowest`, and `aggregation` operations when calculating grades. For me, it's not clear how one controls the order of these operations. Is it just the order they are listed in the YAML for each category? I recommend making this more explicit in the documentation.
