### 2025-10-16 nemogb-dev/nemogb-r#125 @jbburant
Separate export of calculated grades
Relates to review for JOSE: https://github.com/openjournals/jose-reviews/issues/303

The `get_grades` function runs a series of operations to process and reconcile the grade and policy files before calculating the component-specific scores and final grades, which are appended to the original grade sheet. I think one nice functionality would be an argument to `get_grades` that allows export of the calculated values in a separate table (e.g., `append = FALSE`). The current performance, I imagine, is meant to make it easy to upload these calculated score back to the LMS, but I does make it difficult to identify which columns are newly calculated and which come from the provided files. This isn't essential functionality, but would be relatively straightforward to implement.
