### 2025-10-16 nemogb-dev/nemogb-r#126 @jbburant
Broken link in excusing assignments vignette
Relates to review for JOSE: https://github.com/openjournals/jose-reviews/issues/303

In the [Excusing Assignments vignette](https://nemogb-dev.github.io/nemogb-r/articles/excusing-assignments.html), the hyperlink back to the policy file vignette is broken in the following sentence:
"Full details on the working of drops are available **[[here]]**."

The provided link points to the wrong domain "nemo**gv**" rather than "nemogb"
