### 2025-10-16 #303 @jbburant
I've now reviewed the package and paper following the checklist, as well as some more in-depth exploration of the repository, past issues, etc. The paper is clear about the need being addressed by the package and, in general, I think `nemogb` looks like a really useful package for easily, flexibly, and reproducibly calculating course grades. The vignettes included with the package are very useful and I particularly appreciate the inclusion of details on building a policy files that reflect syllabi of varying complexity (this step is obviously key to successfully using the package!).

@nikita-jaya I have raised only a couple small issues on the target repository.
@bleds22e In my view, this software and the accompanying paper are in good order and ready for an editorial decision. I recommend acceptance after the issues I've added are considered/addressed.
