### 2025-10-16 29caf5b
adding append parameter to get_grades()
