+++ R/get-grades.R
@@ -10,7 +10,7 @@
 #' @param gs A Gradescope dataframe
 #' @param policy A grading policy file as an R list
 #' @param verbose if FALSE, throws error if no assignments found in gs
-#' @param append if TRUE, returns dataframe with only newly computed categories' scores
+#' @param append if FALSE, returns dataframe with only newly computed categories' scores
 #'
 #' @examples
 #' get_grades(gs = gs_demo, policy = policy_demo, verbose = TRUE)
@@ -21,8 +21,7 @@
 #' @importFrom tibble as_tibble
 #'
 #' @export
-get_grades <- function(gs, policy, verbose = FALSE,
-                       append = FALSE) {
+get_grades <- function(gs, policy, verbose = FALSE, append = TRUE) {
   gs <- gs |>
     process_gs()
 
@@ -35,7 +34,7 @@ get_grades <- function(gs, policy, verbose = FALSE,
     calculate_grades(policy = policy) |>
     tibble::as_tibble()
   
-  if (append){
+  if (!append){
     # id_cols calls all ID columns and all newly generated columns
     # when it's called on a post-grading dataframe
     grades <- grades[, get_id_cols(grades)]
