+++ man/get_grades.Rd
@@ -4,7 +4,7 @@
 \alias{get_grades}
 \title{Get Grades}
 \usage{
-get_grades(gs, policy, verbose = FALSE, append = FALSE)
+get_grades(gs, policy, verbose = FALSE, append = TRUE)
 }
 \arguments{
 \item{gs}{A Gradescope dataframe}
@@ -13,7 +13,7 @@ get_grades(gs, policy, verbose = FALSE, append = FALSE)
 
 \item{verbose}{if FALSE, throws error if no assignments found in gs}
 
-\item{append}{if TRUE, returns dataframe with only newly computed categories' scores}
+\item{append}{if FALSE, returns dataframe with only newly computed categories' scores}
 }
 \value{
 A dataframe of the original Gradescope data with computed categories' scores appended as additional columns
