### 2025-10-29 c538e13
adding tests for append argument in get_grades
