### 2025-11-05 #303 @nikita-jaya
@bleds22e the authors and affiliations are correct, and the release with the latest changes is v1.2.2

I will archive the release on Zenodo and complete the rest of the "Additional Author Tasks" soon!
