### 2025-11-06 #303 @nikita-jaya
@bleds22e
UPDATES:

- Final version: v1.2.2
- Zenodo DOI: 10.5281/zenodo.17546946
- All other "Additional Author Tasks" are completed

Let me know if there's anything else!
