### 2025-11-19 8eb017e
Refine language and clarify course grading details
