### 2025-11-19 #303 @bleds22e
@nikita-jaya I made a PR with a few minor edits in the manuscript. Once you review/merge that, I can continue.
