### 2025-11-19 #303 @nikita-jaya
@bleds22e All changes look fine, except in a few instances you have generalized "Gradescope or Canvas" to "Gradescope or another LMS"... our workflow expects that the inputted .csv file is formatted based on grades that are Gradescope exports or grades that are Canvas exports, so this generalization would be inaccurate.

Would you be able to change this? And then, I can merge it!
