### 2025-11-19 49e05ca
remove LMS generalizations
