+++ reference/index.md
@@ -1,0 +1,91 @@
+# Package index
+
+## Gradescope Data Functions
+
+These functions are to read in and process data from a grading platform
+such as Gradescope.
+
+- [`read_files()`](https://github.com/nemogb-dev/nemogb-r/reference/read_files.md)
+  : Read in File(s) from Grading Platform(s)
+- [`process_gs()`](https://github.com/nemogb-dev/nemogb-r/reference/process_gs.md)
+  : Process Gradescope Data
+- [`check_data_format()`](https://github.com/nemogb-dev/nemogb-r/reference/check_data_format.md)
+  : Check Formatting of Grades Data
+- [`get_id_cols()`](https://github.com/nemogb-dev/nemogb-r/reference/get_id_cols.md)
+  : Get the ID Columns for Gradescope Data
+- [`get_assignments()`](https://github.com/nemogb-dev/nemogb-r/reference/get_assignments.md)
+  : Get the Assignment Names for Gradescope Data
+- [`determine_grade_source()`](https://github.com/nemogb-dev/nemogb-r/reference/determine_grade_source.md)
+  : Predict the source of grade dataframe.
+- [`read_gs()`](https://github.com/nemogb-dev/nemogb-r/reference/read_gs.md)
+  : Read Gradescope CSV File
+
+## Policy File Functions
+
+These functions are to read in and process the policy file.
+
+- [`read_policy()`](https://github.com/nemogb-dev/nemogb-r/reference/read_policy.md)
+  : Read Grading Policy .yml
+- [`process_policy()`](https://github.com/nemogb-dev/nemogb-r/reference/process_policy.md)
+  : Process Policy File
+- [`reconcile_policy_with_gs()`](https://github.com/nemogb-dev/nemogb-r/reference/reconcile_policy_with_gs.md)
+  : Reconcile Policy File with Gradescope Data
+- [`flatten_policy()`](https://github.com/nemogb-dev/nemogb-r/reference/flatten_policy.md)
+  : Flatten Policy File ("Un-nest" the Nested Structure of Policy File)
+- [`get_categories()`](https://github.com/nemogb-dev/nemogb-r/reference/get_categories.md)
+  : Get the Category Names for Policy
+- [`get_nested_assignments()`](https://github.com/nemogb-dev/nemogb-r/reference/get_nested_assignments.md)
+  : Get the Nested Assignments for Category
+
+## Grading Functions
+
+These functions are to grade the processed Gradescope data based on the
+criteria in the processed policy file
+
+- [`get_grades()`](https://github.com/nemogb-dev/nemogb-r/reference/get_grades.md)
+  : Get Grades
+- [`calculate_grades()`](https://github.com/nemogb-dev/nemogb-r/reference/calculate_grades.md)
+  : Calculate Grades
+- [`get_category_grade()`](https://github.com/nemogb-dev/nemogb-r/reference/get_category_grade.md)
+  : Calculate A Single Category Grade
+- [`score()`](https://github.com/nemogb-dev/nemogb-r/reference/score.md)
+  [`aggregation()`](https://github.com/nemogb-dev/nemogb-r/reference/score.md)
+  [`lateness()`](https://github.com/nemogb-dev/nemogb-r/reference/score.md)
+  [`drop_n_lowest()`](https://github.com/nemogb-dev/nemogb-r/reference/score.md)
+  [`aggregation_max_pts()`](https://github.com/nemogb-dev/nemogb-r/reference/score.md)
+  [`aggregation_lateness()`](https://github.com/nemogb-dev/nemogb-r/reference/score.md)
[…]
