+++ authors.md
@@ -1,0 +1,29 @@
+# Authors and Citation
+
+## Authors
+
+- **Nikita Jayaprakash**. Author, maintainer.
+
+- **Monika Voutov**. Author.
+
+- **Andrew Bray**. Author.
+
+- **Zachary Turner**. Author.
+
+## Citation
+
+Source:
+[`DESCRIPTION`](https://github.com/nemogb-dev/nemogb-r/blob/main/DESCRIPTION)
+
+Jayaprakash N, Voutov M, Bray A, Turner Z (2025). *nemogb: Calculate
+course grades based on assignment scores from Gradescope or Canvas and a
+description of the course grading policy*. R package version 1.2.0,
+<https://github.com/nemogb-dev/nemogb-r>.
+
+    @Manual{,
+      title = {nemogb: Calculate course grades based on assignment scores from Gradescope or Canvas and a description of the course grading policy},
+      author = {Nikita Jayaprakash and Monika Voutov and Andrew Bray and Zachary Turner},
+      year = {2025},
+      note = {R package version 1.2.0},
+      url = {https://github.com/nemogb-dev/nemogb-r},
+    }
