+++ reference/calculate_grades.md
@@ -1,0 +1,236 @@
+# Calculate Grades
+
+This function calculates all grades based on the policy file. This
+function is called within
+[`get_grades()`](https://github.com/nemogb-dev/nemogb-r/reference/get_grades.md)
+because it requires previous data processing. Note that the example
+below is essentially the functionality of
+[`get_grades()`](https://github.com/nemogb-dev/nemogb-r/reference/get_grades.md).
+
+## Usage
+
+``` r
+calculate_grades(gs, policy)
+```
+
+## Arguments
+
+- gs:
+
+  A Gradescope dataframe
+
+- policy:
+
+  A grading policy file as an R list
+
+## Value
+
+A dataframe of the original Gradescope data with computed categories'
+scores appended as additional columns
+
+## Examples
+
+``` r
+gs <- process_gs(gs = gs_demo)
+policy <- process_policy(policy = policy_demo, verbose = TRUE)
+policy <- reconcile_policy_with_gs(policy = policy, gs = gs_demo, verbose = TRUE)
+calculate_grades(gs = gs, policy = policy)
+#>                  Names                           Email        SID
+#> 1            Dr. Kadin           Dr.Kadin@berkeley.edu 3032412514
+#> 2          Bartley Toy         BartleyToy@berkeley.edu 3032412515
+#> 3           Dr. Dontae          Dr.Dontae@berkeley.edu 3032412516
+#> 4  Narcissus Considine NarcissusConsidine@berkeley.edu 3032412517
+#> 5      Verne Wilderman     VerneWilderman@berkeley.edu 3032412518
+#> 6        Toivo Witting       ToivoWitting@berkeley.edu 3032412519
+#> 7         Tegan Crooks        TeganCrooks@berkeley.edu 3032412520
+#> 8  Tyrus Cremin-Wisozk TyrusCremin-Wisozk@berkeley.edu 3032412521
+#> 9         Sydni Hansen        SydniHansen@berkeley.edu 3032412522
+#> 10        Mrs. Chelsie        Mrs.Chelsie@berkeley.edu 3032412523
+#>           Sections Lab 1 Lab 1 - Max Points Lab 1 - Submission Time
+#> 1  fake-class-2023   0.9                 10                       7
+#> 2  fake-class-2023   0.5                 10                       5
+#> 3  fake-class-2023   1.0                 10                       2
+#> 4  fake-class-2023   1.0                 10                      10
+#> 5  fake-class-2023   0.8                 10                       4
+#> 6  fake-class-2023   0.9                 10                       6
+#> 7  fake-class-2023   0.8                 10                       9
+#> 8  fake-class-2023   0.2                 10                       3
+#> 9  fake-class-2023   0.8                 10                       1
+#> 10 fake-class-2023   0.5                 10                       8
+#>    Lab 1 - Lateness (H:M:S) Lab 2 Lab 2 - Max Points Lab 2 - Submission Time
+#> 1                  00:00:00  0.25                 20                      10
+#> 2                  00:00:00  1.00                 20                       6
+#> 3                  00:00:00  0.60                 20                       7
+#> 4                  00:00:00  0.05                 20                       1
+#> 5                  00:00:00  0.50                 20                       3
+#> 6                  00:00:00  0.10                 20                       5
+#> 7                  11:25:51  0.55                 20                       9
+#> 8                  00:00:00  1.00                 20                       2
+#> 9                  00:00:00  0.15                 20                       8
+#> 10                 00:34:45  0.95                 20                       4
+#>    Lab 2 - Lateness (H:M:S)      Lab 3 Lab 3 - Max Points
+#> 1                  47:55:53 0.46666667                 15
+#> 2                  00:00:00 0.06666667                 15
+#> 3                  00:00:00 1.00000000                 15
[…]
