### 2025-11-19 #303 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
