### 2025-12-09 #303 @labarba
While you do that, please merge this for caps-protect in titles:
https://github.com/nemogb-dev/nemogb-r/pull/132
