### 2025-12-09 #303 @labarba
The Summary starts with "NemoGB (Jayaprakash* et al., 2024) "... I'm confused about the citation. The reference title is "NemoGB web app" and it goes to a GitHub repo that is not the repo of this submission, but the language of the summary sounds like they are the same thing. Is this a circular citation?
