### 2025-12-09 #303 @nikita-jaya
@labarba - The metadata for the Zenodo deposit is updated (Silas Santini has been updated), and the cap-protect PR has been merged.

There are two separate citations for "NemoGB Package", which is what this paper is addressing, and "NemoGB Web App", which is a separate repo and is a Shiny UI that uses this package.

Let me know if there's anything else.
