### 2025-12-09 #303 @labarba
small typo here: https://github.com/nemogb-dev/nemogb-r/pull/133
