### 2025-12-09 #303 @labarba
can you check if the Summary text is clear about the different packages? ... I was confused
