### 2025-12-09 #303 @labarba
Also, looks like author ORCIDs are gone now on the [archive](https://zenodo.org/records/17546946).
