### 2025-12-10 #303 @nikita-jaya
@labarba Sorry about that, the Zenodo archive should now have the same title as the paper and the author ORCIDs are added back.

The PR with the typo is pushed as well.

I believe the current Summary text is accurate because it references the NemoGB package, which is what this paper is about. The NemoGB App (which is a separate repo) is only mentioned later in the Statement of Need section.

Let me know if I should change anything else.
