### 2025-12-10 #303 @labarba
> I believe the current Summary text is accurate because it references the NemoGB package, which is what this paper is about.

This is what confuses me: the paper is about the NemoGB package, and you start the paper with a reference to the same package, so it's circular. The _paper itself_ is the reference and the repo is linked on the metadata and margin, so you can just remove this citation.
