### 2025-12-10 1e2f9f8
Update paper.md
