### 2025-12-10 #303 @nikita-jaya
@labarba Oh, I see what you mean now. I've removed the circular citation (both from the paper and the paper.bib).

Thank you for your feedback so far!
