### 2025-12-11 #303 @labarba
Congratulations @nikita-jaya – your JOSE paper is published! 🚀

Tremendous thanks to our Editor: @bleds22e and Reviewers: @lwjohnst86, @jbburant — your contribution makes this possible 🙏
