### 2025-12-11 #303 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
