### 2025-12-11 #303 @bleds22e
Congrats, @nikita-jaya!

@labarba thanks for catching some of the things I missed! I didn't realize the Zenodo and JOSE authors needed to be the same (which seems pretty obvious, in retrospect...).
